A function that converts raw diff into a convenient form for pprint.
APPLY-MODIFIEDNODE-CONVERTERS converts a diff with reference nodes and
lost nodes (``raw diff'') into another form that is easy to handle by
pretty printing step.

The returned form is then processed by pretty printer. Be careful that
the returned form is still a form and not yet pretty printed.

In my plan, another facility that defines APPLY-MODIFIEDNODE-CONVERTERS
will be implemented because I think that more complexed predicator will
make it possible to offer more detailed diffs. As a consequence, it
determines the goodness of result. Therefore, I think we should offer a
facility that can add a predicator and converter pair in an easy way.
While I think too many categories of information in diff harms the
readability of pretty printed diff, I'm not sure if the diff only shows
add/remove is the best one.

More precisely, MERGE-LOST should also be some kind of pluggable or
programmable something because it now offers only add/remove
information. I think replace can be detected from this information, but
what I'm discussing here is about the general facility. In that general
facility, MERGE-LOST will be used to merge not only LOST nodes but also
several kinds of nodes, and that result is called ``raw diff'' for
example.

I think this suite of facilities become a system for diff generator.