[Bugfix] Special treatment for the case that replace whole BASE code.
It is still perfect. The only LOST subtree detected is (LOST 0) if the
whole BASE code is replaced. In such case we need an additional pair of
parenthesis that wraps up (LOST 0) and the new node that replaces BASE
code. I have for now no way to eliminate that otiose parenthesis.