A more powerful abstraction called APPLY-CONVERTERS is introduced.
A converter is a 3 parameter function. It takes node, route, and
codelet. A converter should return a codelet.  Currently, node and
codelet are completely same for new nodes because there is no special
expression format for new node in diff. In order to achieve a suite a
certain consistensy, it is good to introduce a special data structure
for new node, too.  Moreover, nodes can be expressed as a class.