[Bugfix] DIFF must return refmark and lostmark.
From conceptual viewpoint, it is a kind of bug. There is no use of
returned diff if the refmark and lostmark are not given because we
cannot tell which node of the diff is lost node or referenced nodes.
Therefore, another possible fix is to make the parameters be required
parameters.