[Bugfix] Unreachable orphan nodes were failed to merge into rawdiff.
An orphan node means that a node whose parnet is lost. If an inner node
is lost, then all the child nodes become orphan nodes. A node can move
to another place and its siblings (and their children) can become
unreachable because the parent of the node is NOT considered as LOST
nodes -- if one or more descendant node appears, then that (inner) node
is not LOST. I overlooked this point.

Let me show you an example:

    (a (x y)) -> (a z y)

In this case, X is an unreachable orphan node if allowed-distance is
equal to or larger than 2. The inner node whose raw route is (1 0) in
the former code is not considered lost because its descendant Y appears
in the latter.

It is worth to mention that, unreachable orphan nodes never appear if
allowed-distance is 0.

There are 2 possible stories that produce unreachable orphan nodes.
First, an inner node that has unreachable orphan nodes is replaced by
other NEW or REF node. Above example is this case. This case is
processed by the first clause of COND macro in the fixed MERGE-LOST function.
Second, an inner node that has unreachable orphan nodes is disappeared
because the number of node in the same level is reduced. Here is an
example:

    (a b (c d)) -> (a c)

This case is handled by the last clause of COND macro.

The implementation is now too noisy. Especially, using the idea of depth
is not good in this function because it concerns nodes among 2 levels.