RDIFF% is revised.
The data structure called REFDIFF is now omitted and RDIFF% plays a main
role of building RAWDIFF. Although the rawdiff construction algorithm of
RAWDIFF is achieved, the code is still dirty for some points. I'm gonna
revise them immediately, but let me explain them.

First, the implementation of booking object is not proper. CITED
bookings need to hold both of its original place and the place who cites
it. But other bookings, LOST and REF, needs only one. LOST holds its
original place and REF holds the place of node it refers.
The another field named CONT seems similar, but the situation is more
complicated. CONT holds the continuation process in order to re-book on
cancel and it was originally introduced for REF nodes. It does not
requires any argument in REF bookings case, and  quite easy to
construct.

    (lambda () next-level)

I realized that LOST bookings also need a sort recalculation in order to
avoid unreachable orphan nodes. Ideally speaking, this recalculation
should be wrapped up as a lambda expression and should be saved into
CONT but the lambda expression is build when it is needed and evaluated.
Although I don't think the latter approach is bad, the same concept,
recalculation, should be handled in a same way.
 I think the RDIFF% function should be divided into 2 functions. One is
the part which uses TRAVERSE/ROUTE macro, the main body of current
RDIFF%, and the LABELS part of current RDIFF% will be another. The
LABELS part is currently highly depending on modbook and basebook, but
I guess it is easy to modify them to be independent of them. The
function just returns the new booking objects, and the callee side (=
RDIFF% side) will push them into mod/base books. The order of objects in
these books is not important at all.

Second, I have to consider about the structure to implement the RAWDIFF,
each leaf of RAWDIFF especially. The leaf is expressed as a list for
now. It may be difficult to distinguish between NEW nodes and bookings.
Anyhow, we need to put multiple booking objects on one place (= route)
in some cases.
 I have not yet considered enough about this problem.