DIFF functions is implemented.
I finally decided to return 2 trees as outputs of diff. One is called
reftree which is composed of REF nodes and NEW nodes and the another is
called losttree which is composed of LOST nodes and CITED nodes.

And 3 of these 4 nodes, REF/LOST/CITED nodes seems improper. The booking
object seems properer to compose reftrees and losttrees. I already have
an algorithm to construct a losttree from a basebook, and the completely
same algorithm is applicable to construct a reftree from a modbook. The
constructed tree will never have a booking object with child booking
objects because such a nested structure is a kind of hotpatch introduced
on cancel.

And about NEW nodes, I think I should introduce a dedicated structure.