Rawdiff construction and book construction are completely divided.
The check-in part made significantly easy because the rawdiff is no more
needed to hold nested function object. The canceled REF booking must
hold the recalculated booking situation of its children in the previous
algorithm. It was initially designed to construct both of reftree and
losttree without traverse the modified and base codes multiple times.
However, I concluded that losttree is impossible to construct in that
way.

The terminology has been changed. Rawdiff is a tree that is composed of
booking objects.  Rawdiff is now literally ``raw'', in the meaning that
it is available only on the memory. And the output of DIFF function is a
format which can be put into a file. It is composed of marks and routes.
Rawdiff is more useful to make the view of diff because each object has
the code itself.

In order to make reftree a rawdiff defined above, NEW booking is
expressed as a functon object likewise as other booking objects.